workspace manifest factory implement filter function
