+++ scopes/dependencies/dependency-resolver/manifest/index.ts
@@ -1,2 +1,3 @@
 export { ComponentManifest } from './component-manifest';
 export { WorkspaceManifest } from './workspace-manifest';
+export { WorkspaceManifestFactory } from './workspace-manifest-factory';
